refactor: Utilities directory reorganization
