--- src/poetryData.js
@@ -1,9 +1,0 @@
-const commonWords = [
-  {type: 'time', words: ['time','life', 'years','long', 'day', 'Time', 'Life', 'Years','Long', 'Day']},
-  {type: 'love', words:['love','light','air','white','heart','Love','Light','Air','White','Heart']},
-  {type: 'spooky', words: ['night','dead','dark','left', 'world', 'Night','Dead', 'Dark', 'Left', 'World']},
-  {type: 'body', words: ['hand','body','face', 'eyes', 'man', 'Hand','Body','Face', 'Eyes', 'Man']},
-  {type: 'nature', words: ['sky','sun','wind','god','water', 'Sky','Sun','Wind','God','Water']}
-]
-
-export default commonWords;
